killercoda: check if a kubeconfig exists before attempting to symlink it

+++ killercoda/install.sh
@@ -22,9 +22,11 @@ function xk6-disruptor() {
 
   echo "xk6-disruptor $XK6_DISRUPTOR_VERSION installed to $XK6_DISRUPTOR_INSTALL_PATH"
 
-  echo "Symlinking k3s kubeconfig to default path"
-  mkdir -p ~/.kube
-  ln -s /etc/rancher/k3s/k3s.yaml ~/.kube/config
+  if [ ! -e ~/.kube/config ]; then
+    echo "Symlinking k3s kubeconfig to default path"
+    mkdir -p ~/.kube
+    ln -s /etc/rancher/k3s/k3s.yaml ~/.kube/config
+  fi
 }
 
 function sock-shop() {
